Adding Options and Safety
This commit adds the ability to specify what you want to delete.
You can combine the flags to select specific elements. Or just
use the -a flag to clear everythign.

Additionally it adds a final check to make sure you are sure that you
want to delete things. You must type 'YES' (case sensitive) to
continue with removing those items from the system.